[SSH Connect] Connect to Jupyter Notebook using port forwarding (#6)
* Added connect and stop commands

* Added functionality for adding, listing and deleting security groups and list instances

* Added functionality for connecting to Notebook server

* Added script to handle ssh tunneling

* Added paramiko and requests to dependencies

* Clean up

* Made exception to catch a parameter

* ClientError as exception for retry

* Open notebook in browser automatically

* Bump version

+++ bokchoi/ec2.py
@@ -2,10 +2,10 @@
 """
 Class which can be used to deploy and run EC2 spot instances
 """
-import os
 from base64 import b64encode
+import os
 
-from bokchoi import common
+from bokchoi import common, ssh
 
 USER_DATA = """#!/bin/bash
 
@@ -23,11 +23,25 @@ python3 -c "import zipfile; zf = zipfile.ZipFile('/tmp/{package}'); zf.extractal
 curl -sS https://bootstrap.pypa.io/get-pip.py | sudo python3
 [ -f /tmp/requirements.txt ] && pip3 install -r /tmp/requirements.txt
 
-# Run app
-cd /tmp
-python3 -c "import {app}; {app}.{entry}();"
-aws s3 cp /var/log/cloud-init-output.log s3://{bucket}/cloud-init-output.log
-shutdown -h now
+if [ "{notebook}" = "True" ]
+then
+    #Add public key
+    echo "ssh-rsa {public_key}" >> /home/ubuntu/.ssh/authorized_keys
+    #Install Jupyter
+    pip3 install jupyter
+    pip3 install tornado==4.5.2
+    echo "c.NotebookApp.token = u''" >> ~/.jupyter/jupyter_notebook_config.py
+    jupyter notebook --no-browser --allow-root --ip=0.0.0.0 --port=8888 --NotebookApp.token=
+else
+    # Run app
+    cd /tmp
+    python3 -c "import {app}; {app}.{entry}();"
+    aws s3 cp /var/log/cloud-init-output.log s3://{bucket}/cloud-init-output.log
+    if [ "{shutdown}" = "True" ]
+    then
+        shutdown -h now
+    fi
+fi
 """
 
 DEFAULT_TRUST_POLICY = """{
@@ -71,6 +85,8 @@ class EC2(object):
         self.launch_config = settings.get('EC2')
         self.custom_policy = settings.get('CustomPolicy')
 
+        self.shutdown = settings.get('Shutdown', True)
+
         aws_account_id = common.get_aws_account_id()
         self.project_id = common.create_project_id(project, aws_account_id)
         self.package_name = 'bokchoi-' + self.project_name + '.zip'
@@ -90,6 +106,15 @@ class EC2(object):
 
         self.create_default_role_and_profile(policies)
 
+        if self.settings.get('Notebook'):
+            common.create_security_group(self.project_id + '-default'
+                                         , self.project_id
+                                         , {'CidrIp': common.get_my_ip() + '/32'
+                                            , 'FromPort': 22
+                                            , 'ToPort': 22
+                                            , 'IpProtocol': 'tcp'}
+                                         )
+
         if self.settings.get('Schedule'):
 
             from bokchoi.scheduler import Scheduler
@@ -104,7 +129,9 @@ class EC2(object):
         """Deletes all policies, users, and instances permanently"""
 
         common.cancel_spot_request(self.project_id, dryrun)
-        common.terminate_instances(self.project_id, dryrun)
+
+        for instance in common.get_instances(self.project_id):
+            common.terminate_instance(instance, dryrun)
 
         common.delete_bucket(self.project_id, dryrun)
 
@@ -116,6 +143,9 @@ class EC2(object):
 
         for role in common.get_roles(self.project_id):
             common.delete_role(role, dryrun)
+
+        for group in common.get_security_groups(self.project_id):
+            common.delete_security_group(group, dryrun)
 
         from bokchoi.scheduler import Scheduler
 
@@ -131,11 +161,32 @@ class EC2(object):
 
         bucket_name = self.project_id
 
-        app, entry = self.entry_point.split('.')
-        user_data = USER_DATA.format(bucket=bucket_name, package=self.package_name, app=app, entry=entry)
+        if self.settings.get('Notebook'):
+            public_key = ssh.get_ssh_keys(self.project_id)
+        else:
+            public_key = ''
+
+        if self.entry_point:
+            app, entry = self.entry_point.split('.')
+        else:
+            app, entry = '', ''
+
+        user_data = USER_DATA.format(bucket=bucket_name
+                                     , package=self.package_name
+                                     , app=app
+                                     , entry=entry
+                                     , public=public_key
+                                     , shutdown=self.shutdown
+                                     , notebook=self.settings.get('Notebook')
+                                     , public_key=public_key)
+
         self.launch_config['LaunchSpecification']['UserData'] = b64encode(user_data.encode('ascii')).decode('ascii')
 
         self.launch_config['LaunchSpecification']['IamInstanceProfile'] = {'Name': self.project_id + '-default-role'}
+
+        if self.settings.get('Notebook'):
+            security_group = next(common.get_security_groups(self.project_id, self.project_id + '-default'))
+            self.launch_config['LaunchSpecification']['SecurityGroupIds'] = [security_group.group_id]
 
         common.request_spot_instances(self.project_id, self.launch_config)
 
@@ -166,3 +217,18 @@ class EC2(object):
             policies.append(next(common.get_policies(custom_policy_name)))
 
         return policies
+
+    def connect(self):
+        """Set up port forwarding for ipython notebook server"""
+        instance = next(common.get_instances(self.project_id))
+        instance_ip = instance.public_ip_address or instance.private_ip_address
+        key_file = os.path.expanduser('~/.ssh/' + self.project_id)
+        ssh.forward(8888, instance_ip, 'ubuntu', key_file)
+
+    def stop(self):
+        """Stop all running instances"""
+        dryrun = False
+        common.cancel_spot_request(self.project_id, dryrun)
+
+        for instance in common.get_instances(self.project_id):
+            common.terminate_instance(instance, dryrun)
